Upgraded kForth environment to 1.6.6; added error check in Expression compiler.
 Changes to be committed:
	modified:   src/AlgebraCompiler.cpp
	modified:   src/CPlotWindow.cpp
	modified:   src/ForthCompiler.cpp
	modified:   src/ForthCompiler.h
	modified:   src/ForthVM.cpp
	modified:   src/ForthVM.h
	modified:   src/ForthWords.h
	modified:   src/Makefile
	modified:   src/fbc.h
	modified:   src/vm32-common.s
	deleted:    src/vm32-fast.s
	modified:   src/vm32.s
	modified:   src/vmc.c
	modified:   src/xyplot.cpp
	modified:   src/xyplot.h

+++ src/xyplot.h
@@ -27,7 +27,7 @@ void LoadInitializationFile();
 vector<char*> GetStartupFileList (int, char**);
 void InitForthInterface();
 int LoadForthFile(char*);
-int ExecuteForthExpression (char*, ostringstream*, int*);
+int ExecuteForthExpression (char*, ostringstream*, long int*);
 
 struct IfcFuncTemplate
 {
